fix(weight-free,dtype): respect model config dtype throughout export pipeline
- core.py: use config.dtype (not hardcoded float32) for meta model cast and
  checkpoint conversion; getattr fallback handles dtype=None configs
- core.py: propagate enable_proxy flag through _build_meta_qeff_model
- cache_utils.py: use v_out.dtype instead of torch.float32 for invalid-mask
  zeros_like in 5 KV cache update paths — avoids fp32↔bf16 dtype mismatch
- modeling_llama.py: use attn_weights.dtype for masked_fill and softmax
  upcast instead of hardcoded float32
- export_compile_infer.py: add --dtype flag (default float16) to control
  model weight dtype; add enable_proxy=False
- modeling_qeff.py: skip SplitTensorsTransform for weight-free export

+++ QEfficient/transformers/cache_utils.py
@@ -213,7 +213,7 @@ class QEffDynamicLayer(CacheLayerMixin):
             v_out = ctx_gather_interface(v_out, ctx_indices, ctx_len)
 
         invalid_mask = _match_invalid_mask(invalid_mask, v_out.shape[-2])
-        v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=torch.float32), v_out)
+        v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=v_out.dtype), v_out)
 
     def read_only_blockedKV(self, start_index, end_index, cache_kwargs):
         """
@@ -266,7 +266,7 @@ class QEffDynamicLayer(CacheLayerMixin):
             )
 
         invalid_mask = _match_invalid_mask(invalid_mask, v_out.shape[-2])
-        v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=torch.float32), v_out)
+        v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=v_out.dtype), v_out)
         return k_out, v_out
 
     def write_only(self, key_states, value_states, cache_kwargs):
@@ -388,7 +388,7 @@ class QEffDynamicLayer(CacheLayerMixin):
                 v_out = ctx_gather_interface(v_out, ctx_indices, ctx_len)
 
             invalid_mask = _match_invalid_mask(invalid_mask, v_out.shape[-2])
-            v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=torch.float32), v_out)
+            v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=v_out.dtype), v_out)
 
         return k_out, v_out
 
@@ -468,7 +468,7 @@ class QEffDynamicLayer(CacheLayerMixin):
                 k_out = ctx_gather_3d_interface(k_out, ctx_indices)
                 v_out = ctx_gather_3d_interface(v_out, ctx_indices)
 
-            v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=torch.float32), v_out)
+            v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=v_out.dtype), v_out)
 
         return k_out, v_out
 
@@ -942,7 +942,7 @@ class QEffHybridCache(HybridCache):
             )
             k_out = ctx_gather_interface(k_out, final_indices, ctx_len)
             v_out = ctx_gather_interface(v_out, final_indices, ctx_len)
-            ctx_v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=torch.float32), v_out)
+            ctx_v_out = torch.where(invalid_mask.unsqueeze(-1), torch.zeros_like(v_out, dtype=v_out.dtype), v_out)
             v_out = torch.where((is_sliding_layer & (position_ids.max() >= (layer_ctx_len - 1))), v_out, ctx_v_out)
         return k_out, v_out
